refactor: extract VideoDownloadMenu component from VideoDialog (3c)
The toolbar download menu (files, subtitle .vtt, jw.org link) moves to
components/player/VideoDownloadMenu.vue together with the jwOrgUrl
computed it renders.
