+++ app/components/VideoDialog.vue
@@ -12,42 +12,7 @@
         </v-toolbar-title>
 
         <template #append>
-          <v-menu location="bottom end" transition="slide-y-transition">
-            <template #activator="{ props }">
-              <v-btn icon v-bind="props">
-                <v-icon>mdi-download</v-icon>
-              </v-btn>
-            </template>
-
-            <v-list density="compact">
-              <v-list-subheader>{{ store.translations.btnDownload ?? 'Download' }}</v-list-subheader>
-
-              <v-list-item
-                v-for="file in downloadableFiles(videoMedia)"
-                :key="file.checksum"
-                :href="file.progressiveDownloadURL"
-                prepend-icon="mdi-download"
-                :title="`${file.label} (${Math.floor(file.filesize / 1048576)} MB)`"
-              />
-
-              <v-list-item
-                :disabled="!subtitleUrl"
-                :href="subtitleUrl ?? undefined"
-                prepend-icon="mdi-download"
-                :title="`${store.translations.hdgSubtitles} (.vtt)`"
-              />
-
-              <v-divider class="my-1" />
-
-              <v-list-item
-                :disabled="!jwOrgUrl"
-                :href="jwOrgUrl ?? undefined"
-                prepend-icon="mdi-open-in-new"
-                target="_blank"
-                :title="store.translations.lnkHome"
-              />
-            </v-list>
-          </v-menu>
+          <PlayerVideoDownloadMenu :subtitle-url="subtitleUrl" :video-media="videoMedia" />
 
           <v-btn icon @click="dialog = false">
             <v-icon>mdi-close</v-icon>
@@ -133,7 +98,6 @@
 </template>
 
 <script setup lang="ts">
-import type { Video } from '~/types';
 import { useDisplay } from 'vuetify';
 
 const store = useAppStore();
@@ -187,15 +151,6 @@ const subtitleLanguage = computed({
 const videoPoster = computed(
   () => (store.selectedVideo ?? videoMedia.value)?.images.wss.lg,
 );
-
-const jwOrgUrl = computed(() => {
-  const video = store.selectedVideo ?? videoMedia.value;
-  if (!video) {
-    return null;
-  }
-  const { locale } = store.getSiteLanguage!;
-  return `https://www.jw.org/finder?locale=${locale}&category=${video.primaryCategory}&lank=${video.languageAgnosticNaturalKey}`;
-});
 
 const availableLanguages = computed(() => {
   if (!store.selectedVideo) {
